upgrades for DFG v0.5.3
